Remove DeviceInfo::created
This was removed in https://github.com/signalapp/Signal-Server/commit/31d6ac71a4b5d4f929a4f20468a0b9a7336c465f.
There seems to be a replacement createdAtCiphertext available.

I also replaced device IDs when unlinking the device from i64 to u8, as
that is also the type used inside DeviceInfo.

+++ src/push_service/linking.rs
@@ -82,7 +82,7 @@ impl PushService {
         .map_err(Into::into)
     }
 
-    pub async fn unlink_device(&mut self, id: i64) -> Result<(), ServiceError> {
+    pub async fn unlink_device(&mut self, id: u8) -> Result<(), ServiceError> {
         self.request(
             Method::DELETE,
             Endpoint::service(format!("/v1/devices/{}", id)),
